add logs, move url, change alert description

+++ src/murfey/server/api/session_info.py
@@ -277,7 +277,7 @@ def update_current_gain_ref(
     db.commit()
 
 
-# Methods for turning alerts on and off
+# Endpoints for turning alerts on and off
 
 
 @router.get("/silences/{instrument_name}")
@@ -288,11 +288,18 @@ def get_silences(instrument_name: MurfeyInstrumentName):
         return None
     machine_config = all_configs[instrument_name]
     alertmanager_url = machine_config.alertmanager_url
-    silences = requests.get(
+    if alertmanager_url == "":
+        logger.warning(f"alertmanager_url not set for {sanitise(instrument_name)}")
+        return None
+    response = requests.get(
         f"{alertmanager_url}/api/v2/silences?filter=microscope={sanitise(instrument_name)}"
     )
+    if response.status_code != 200:
+        logger.warning(
+            f"Tried to get silences for {sanitise(instrument_name)}, but received status {response.status_code} from alertmanager API"
+        )
     active_silences = []
-    for silence in silences.json():
+    for silence in response.json():
         if silence["status"]["state"] == "active":
             active_silences.append(silence)
     return active_silences
@@ -302,6 +309,9 @@ def get_silences(instrument_name: MurfeyInstrumentName):
 def create_silence(instrument_name: MurfeyInstrumentName, end_time: datetime):
     machine_config = machine_info_by_instrument(instrument_name)
     alertmanager_url = machine_config.alertmanager_url
+    if alertmanager_url == "":
+        logger.warning(f"alertmanager_url not set for {sanitise(instrument_name)}")
+        return None
     start_time = datetime.now().astimezone().isoformat()
     end_time_str = end_time.astimezone().isoformat()
     silence_json = {
@@ -309,13 +319,17 @@ def create_silence(instrument_name: MurfeyInstrumentName, end_time: datetime):
             {"name": "microscope", "value": instrument_name, "isRegex": False}
         ],
         "createdBy": "murfey",
-        "annotations": {"description": "Test"},
+        "annotations": {"description": f"Silence for alerts on {instrument_name}"},
         "comment": "silence created from murfey",
         "status": {"state": "active"},
         "startsAt": str(start_time),
         "endsAt": str(end_time_str),
     }
     response = requests.post(f"{alertmanager_url}/api/v2/silences", json=silence_json)
+    if response.status_code != 200:
+        logger.warning(
+            f"Tried to create silence for {sanitise(instrument_name)}, but received status {response.status_code} from alertmanager API"
+        )
     return JSONResponse(
         status_code=response.status_code, content=response.json()
     )  # return a response with same data and code as from alertmanager
@@ -327,12 +341,19 @@ def create_silence(instrument_name: MurfeyInstrumentName, end_time: datetime):
 def delete_silences(instrument_name: MurfeyInstrumentName):
     machine_config = machine_info_by_instrument(instrument_name)
     alertmanager_url = machine_config.alertmanager_url
+    if alertmanager_url == "":
+        logger.warning(f"alertmanager_url not set for {sanitise(instrument_name)}")
+        return None
     silences = get_silences(instrument_name)
     if len(silences) == 0:
         return None
     for silence in silences:
         id = silence["id"]
         response = requests.delete(f"{alertmanager_url}/api/v2/silence/{id}")
+        if response.status_code != 200:
+            logger.warning(
+                f"Tried to delete silence {id} for {sanitise(instrument_name)}, but received status {response.status_code} from alertmanager API"
+            )
     return response  # returns final response in loop
